refactor(settings): split Controller and Service into Public and Admin API
- split for easier management
