+++ src/main/java/de/fsr/mariokart_backend/settings/controller/admin/AdminSettingsDeleteController.java
@@ -1,26 +1,28 @@
-package de.fsr.mariokart_backend.settings.controller;
+package de.fsr.mariokart_backend.settings.controller.admin;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.server.ResponseStatusException;
 
+import de.fsr.mariokart_backend.controller.annotation.ApiController;
+import de.fsr.mariokart_backend.controller.annotation.ApiType;
+import de.fsr.mariokart_backend.controller.annotation.ControllerType;
 import de.fsr.mariokart_backend.exception.RoundsAlreadyExistsException;
-import de.fsr.mariokart_backend.settings.service.SettingsDeleteService;
+import de.fsr.mariokart_backend.settings.service.admin.AdminSettingsDeleteService;
 import lombok.AllArgsConstructor;
 
 @RestController
-@RequestMapping("/settings")
 @AllArgsConstructor
-public class SettingsDeleteController {
+@ApiController(apiType = ApiType.ADMIN, controllerType = ControllerType.SETTINGS)
+public class AdminSettingsDeleteController {
 
-    private final SettingsDeleteService settingsDeleteService;
+    private final AdminSettingsDeleteService adminSettingsDeleteService;
 
     @DeleteMapping("/reset")
     public void reset() {
         try {
-            settingsDeleteService.reset();
+            adminSettingsDeleteService.reset();
         } catch (RoundsAlreadyExistsException e) {
             throw new ResponseStatusException(HttpStatus.CONFLICT, e.getMessage());
         }
