+++ src/main/java/de/fsr/mariokart_backend/settings/controller/admin/AdminSettingsUpdateController.java
@@ -1,30 +1,32 @@
-package de.fsr.mariokart_backend.settings.controller;
+package de.fsr.mariokart_backend.settings.controller.admin;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 
+import de.fsr.mariokart_backend.controller.annotation.ApiController;
+import de.fsr.mariokart_backend.controller.annotation.ApiType;
+import de.fsr.mariokart_backend.controller.annotation.ControllerType;
 import de.fsr.mariokart_backend.exception.RoundsAlreadyExistsException;
 import de.fsr.mariokart_backend.settings.model.dto.TournamentDTO;
-import de.fsr.mariokart_backend.settings.service.SettingsUpdateService;
+import de.fsr.mariokart_backend.settings.service.admin.AdminSettingsUpdateService;
 import lombok.AllArgsConstructor;
 
 @RestController
-@RequestMapping("/settings")
 @AllArgsConstructor
-public class SettingsUpdateController {
+@ApiController(apiType = ApiType.ADMIN, controllerType = ControllerType.SETTINGS)
+public class AdminSettingsUpdateController {
 
-    private final SettingsUpdateService settingsUpdateService;
+    private final AdminSettingsUpdateService adminSettingsUpdateService;
 
     @PutMapping
     @ResponseBody
     public ResponseEntity<TournamentDTO> updateSettings(@RequestBody TournamentDTO tournamentDTO) {
         try {
-            return ResponseEntity.ok(settingsUpdateService.updateSettings(tournamentDTO));
+            return ResponseEntity.ok(adminSettingsUpdateService.updateSettings(tournamentDTO));
         } catch (IllegalStateException e) {
             return ResponseEntity.notFound().build();
         } catch (RoundsAlreadyExistsException e) {
