+++ src/main/java/de/fsr/mariokart_backend/settings/model/Tournament.java
@@ -16,8 +16,6 @@ import lombok.Setter;
 @NoArgsConstructor
 @AllArgsConstructor
 @Table(name = "tournament")
-// @JsonIdentityInfo(generator = ObjectIdGenerators.PropertyGenerator.class,
-// property = "id")
 public class Tournament {
 
     @Id
